Update display of answer key and answer sheet - Add display of AnswerKey object - Change display of AnswerSheet object - Add converter for AnswerKey - Remove unused files

+++ Android/AnswerSheetScorer/app/src/main/java/io/github/stevenalbert/answersheetscorer/dao/AnswerKeyDao.java
@@ -18,7 +18,7 @@ public interface AnswerKeyDao {
     @Insert
     long insert(AnswerKey answerKey);
     @Update
-    long update(AnswerKey answerKey);
+    int update(AnswerKey answerKey);
     @Query("SELECT * FROM answer_key")
     LiveData<List<AnswerKey>> getAllAnswerKeys();
     @Query("SELECT * FROM answer_key WHERE m_code = :mCode")
